+++ test/Utils.spec.ts
@@ -1,21 +1,11 @@
 import { promisify } from '../utils/index';
 import path from 'path';
 import {
-  datetimeToHybrids,
-  hybridtsToUnixtime,
-  unixtimeToHybridts,
-  formatAddress,
-  stringToBase64,
-  parseToKeyValue,
-  formatKeyValueData,
-  formatNumberPrecision,
   getGRPCService,
   getAuthInterceptor,
   checkTimeParam,
   assignTypeParams,
   checkCollectionFields,
-  parseTimeToken,
-  checkRoundDecimal,
 } from '../utils';
 import { ERROR_REASONS, FieldType, DataType } from '../milvus';
 import { generateInsertData } from '../utils/test';
@@ -237,101 +227,6 @@ describe(`Utils`, () => {
     expect(result).toEqual('result');
   });
 
-  it(`hybridtsToUnixtime should success`, async () => {
-    let unixtime = hybridtsToUnixtime('429642767925248000');
-    expect(unixtime).toEqual('1638957092');
-  });
-
-  it(`hybridtsToUnixtime should throw error`, async () => {
-    try {
-      hybridtsToUnixtime(1 as any);
-    } catch (error) {
-      expect(error.message).toContain(ERROR_REASONS.TIMESTAMP_PARAM_CHECK);
-    }
-  });
-
-  it(`unixtimeToHybridts should success`, async () => {
-    let unixtime = unixtimeToHybridts('1638957092');
-    expect(unixtime).toEqual('429642767925248000');
-  });
-
-  it(`unixtimeToHybridts should throw error`, async () => {
-    try {
-      unixtimeToHybridts(1 as any);
-    } catch (error) {
-      expect(error.message).toContain(ERROR_REASONS.TIMESTAMP_PARAM_CHECK);
-    }
-
-    try {
-      unixtimeToHybridts('asd');
-    } catch (error) {
-      expect(error.message).toContain(ERROR_REASONS.TIMESTAMP_PARAM_CHECK);
-    }
-  });
-
-  it(`datetimeToHybrids should success`, async () => {
-    let unixtime = datetimeToHybrids(new Date(1638957092 * 1000));
-    expect(unixtime).toEqual('429642767925248000');
-  });
-
-  it(`datetimeToHybrids should throw error`, async () => {
-    try {
-      datetimeToHybrids(1 as any);
-    } catch (error) {
-      expect(error.message).toContain(ERROR_REASONS.DATE_TYPE_CHECK);
-    }
-  });
-
-  it(`all kinds of url should be supported`, async () => {
-    const port = `80980`;
-    const urlWithHttps = `https://my-url:${port}`;
-    expect(formatAddress(urlWithHttps)).toBe(`my-url:${port}`);
-
-    const urlWithHttp = `https://my-url:${port}`;
-    expect(formatAddress(urlWithHttp)).toBe(`my-url:${port}`);
-
-    const urlWithoutHttp = `my-url`;
-    expect(formatAddress(urlWithoutHttp)).toBe(`my-url:19530`);
-
-    const urlWithoutHttpCustomPort = `my-url:12345`;
-    expect(formatAddress(urlWithoutHttpCustomPort)).toBe(`my-url:12345`);
-
-    const urlWithEmpty = `://my-url`;
-    expect(formatAddress(urlWithEmpty)).toBe(`my-url:19530`);
-
-    const urlWithEmptyCustomPort = `://my-url:12345`;
-    expect(formatAddress(urlWithEmptyCustomPort)).toBe(`my-url:12345`);
-  });
-
-  it(`should convert string to base64 encoding`, () => {
-    const testString = 'hello world, I love milvus';
-    const str = stringToBase64(testString);
-    expect(str.length % 4 == 0 && /^[A-Za-z0-9+/]+[=]{0,2}$/.test(str)).toBe(
-      true
-    );
-  });
-
-  it(`should convert [{key:"row_count",value:4}] to {row_count:4}`, () => {
-    const testValue = [{ key: 'row_count', value: 4 }];
-    const res = formatKeyValueData(testValue, ['row_count']);
-    expect(res).toMatchObject({ row_count: 4 });
-  });
-
-  it(`should convert  {row_count:4} t0 [{key:"row_count",value:4}]`, () => {
-    const testValue = { row_count: 4, b: 3 };
-    const res = parseToKeyValue(testValue);
-    expect(res).toMatchObject([
-      { key: 'row_count', value: 4 },
-      { key: 'b', value: 3 },
-    ]);
-  });
-
-  it(`should convert [{key:"row_count",value:4}] to {row_count:4}`, () => {
-    const testValue = 3.1231241241234124124;
-    const res = formatNumberPrecision(testValue, 3);
-    expect(res).toBe(3.123);
-  });
-
   it(`should return a service client constructor`, () => {
     const protoPath = path.resolve(__dirname, '../proto/proto/milvus.proto');
     const proto = {
@@ -538,36 +433,5 @@ describe(`Utils`, () => {
     const data = generateInsertData(fields, 10);
     expect(data.length).toBe(10);
     expect(typeof data[0].int_field).toBe('number');
-  });
-  test('parses time tokens correctly', () => {
-    expect(parseTimeToken('1s')).toBe(1000);
-    expect(parseTimeToken('2m')).toBe(120000);
-    expect(parseTimeToken('3h')).toBe(10800000);
-    expect(parseTimeToken('4d')).toBe(345600000);
-    expect(parseTimeToken('1w')).toBe(604800000);
-    expect(parseTimeToken('1M')).toBe(2592000000);
-    expect(parseTimeToken('1Y')).toBe(31536000000);
-  });
-
-  test('throws an error for invalid time tokens', () => {
-    expect(() => parseTimeToken('')).toThrow('Invalid time token: ');
-    expect(() => parseTimeToken('1')).toThrow('Invalid time token: 1');
-    expect(() => parseTimeToken('1x')).toThrow('Invalid time token: 1x');
-  });
-
-  it('should not throw an error when round_decimal is undefined', () => {
-    expect(() => checkRoundDecimal(undefined)).not.toThrow();
-  });
-
-  it('should not throw an error when round_decimal is an integer between -1 and 6', () => {
-    expect(() => checkRoundDecimal(0)).not.toThrow();
-    expect(() => checkRoundDecimal(-1)).not.toThrow();
-    expect(() => checkRoundDecimal(6)).not.toThrow();
-  });
-
-  it('should throw an error when round_decimal is not an integer between -1 and 6', () => {
-    expect(() => checkRoundDecimal(1.5)).toThrow();
-    expect(() => checkRoundDecimal(-2)).toThrow();
-    expect(() => checkRoundDecimal(7)).toThrow();
   });
 });
